services/mcp-context/requirements_enhanced.txt
# Enhanced Context MCP with LlamaIndex - Requirements
# Core FastAPI and async support
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.2

# Database drivers
asyncpg==0.29.0
psycopg2-binary==2.9.9

# LlamaIndex core components
llama-index==0.9.30
llama-index-core==0.9.30
llama-index-embeddings-openai==0.1.7
llama-index-llms-openai==0.1.15
llama-index-readers-file==0.1.12
llama-index-node-parser-semantic==0.1.0
llama-index-extractors-keyword==0.1.0
# Removed qdrant vector store - using Weaviate instead

# File processing and readers
PyPDF2==3.0.1
python-docx==1.1.0
unstructured==0.11.2
python-magic==0.4.27

# Vector database
weaviate-client==4.4.1

# OpenAI integration
openai==1.12.0

# Utilities
python-dotenv==1.0.1
python-json-logger==2.0.7
aiofiles==23.2.1

# Optional: Advanced NLP features
nltk==3.8.1
textstat==0.7.3
spacy==3.7.2

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.27.0